Move symbol search out of graphqlbackend
