+++ internal/search/run/symbol.go
@@ -3,20 +3,28 @@ package run
 import (
 	"context"
 	"fmt"
+	"regexp/syntax"
 	"sort"
+	"time"
 
+	"github.com/google/zoekt"
+	zoektquery "github.com/google/zoekt/query"
 	"github.com/neelance/parallel"
 	"github.com/opentracing/opentracing-go/ext"
 	otlog "github.com/opentracing/opentracing-go/log"
+	"github.com/pkg/errors"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
+	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
+	zoektutil "github.com/sourcegraph/sourcegraph/internal/search/zoekt"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
+	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 )
 
@@ -180,4 +188,185 @@ func searchSymbolsInRepo(ctx context.Context, repoRevs *search.RepositoryRevisio
 		return fileMatches[i].Path < fileMatches[j].Path
 	})
 	return fileMatches, err
+}
+
+// indexedSymbols checks to see if Zoekt has indexed symbols information for a
+// repository at a specific commit. If it has it returns the branch name (for
+// use when querying zoekt). Otherwise an empty string is returned.
+func indexedSymbolsBranch(ctx context.Context, repository, commit string) string {
+	z := search.Indexed()
+	if !z.Enabled() {
+		return ""
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, time.Second)
+	defer cancel()
+	set, err := z.ListAll(ctx)
+	if err != nil {
+		return ""
+	}
+
+	repo, ok := set[repository]
+	if !ok || !repo.HasSymbols {
+		return ""
+	}
+
+	for _, branch := range repo.Branches {
+		if branch.Version == commit {
+			return branch.Name
+		}
+	}
+
+	return ""
+}
+
+func searchZoektSymbols(ctx context.Context, repoName types.RepoName, commitID api.CommitID, inputRev *string, branch string, queryString *string, first *int32, includePatterns *[]string) (res []*result.SymbolMatch, err error) {
+	raw := *queryString
+	if raw == "" {
+		raw = ".*"
+	}
+
+	expr, err := syntax.Parse(raw, syntax.ClassNL|syntax.PerlX|syntax.UnicodeGroups)
+	if err != nil {
+		return
+	}
+
+	var query zoektquery.Q
+	if expr.Op == syntax.OpLiteral {
+		query = &zoektquery.Substring{
+			Pattern: string(expr.Rune),
+			Content: true,
+		}
+	} else {
+		query = &zoektquery.Regexp{
+			Regexp:  expr,
+			Content: true,
+		}
+	}
+
+	ands := []zoektquery.Q{
+		&zoektquery.RepoBranches{Set: map[string][]string{
+			string(repoName.Name): {branch},
+		}},
+		&zoektquery.Symbol{Expr: query},
+	}
+	for _, p := range *includePatterns {
+		q, err := zoektutil.FileRe(p, true)
+		if err != nil {
+			return nil, err
+		}
+		ands = append(ands, q)
+	}
+
+	final := zoektquery.Simplify(zoektquery.NewAnd(ands...))
+	match := limitOrDefault(first) + 1
+	resp, err := search.Indexed().Client.Search(ctx, final, &zoekt.SearchOptions{
+		Trace:                  ot.ShouldTrace(ctx),
+		MaxWallTime:            3 * time.Second,
+		ShardMaxMatchCount:     match * 25,
+		TotalMaxMatchCount:     match * 25,
+		ShardMaxImportantMatch: match * 25,
+		TotalMaxImportantMatch: match * 25,
+		MaxDocDisplayCount:     match,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	for _, file := range resp.Files {
+		newFile := &result.File{
+			Repo:     repoName,
+			CommitID: commitID,
+			InputRev: inputRev,
+			Path:     file.FileName,
+		}
+
+		for _, l := range file.LineMatches {
+			if l.FileName {
+				continue
+			}
+
+			for _, m := range l.LineFragments {
+				if m.SymbolInfo == nil {
+					continue
+				}
+
+				res = append(res, &result.SymbolMatch{
+					Symbol: result.Symbol{
+						Name:       m.SymbolInfo.Sym,
+						Kind:       m.SymbolInfo.Kind,
+						Parent:     m.SymbolInfo.Parent,
+						ParentKind: m.SymbolInfo.ParentKind,
+						Path:       file.FileName,
+						Line:       l.LineNumber,
+						Language:   file.Language,
+					},
+					File: newFile,
+				})
+			}
+		}
+	}
+	return
+}
+
+func ComputeSymbols(ctx context.Context, repoName types.RepoName, commitID api.CommitID, inputRev *string, query *string, first *int32, includePatterns *[]string) (res []*result.SymbolMatch, err error) {
+	// TODO(keegancsmith) we should be able to use indexedSearchRequest here
+	// and remove indexedSymbolsBranch.
+	if branch := indexedSymbolsBranch(ctx, string(repoName.Name), string(commitID)); branch != "" {
+		return searchZoektSymbols(ctx, repoName, commitID, inputRev, branch, query, first, includePatterns)
+	}
+
+	ctx, done := context.WithTimeout(ctx, 5*time.Second)
+	defer done()
+	defer func() {
+		if ctx.Err() != nil && len(res) == 0 {
+			err = errors.New("processing symbols is taking longer than expected. Try again in a while")
+		}
+	}()
+	var includePatternsSlice []string
+	if includePatterns != nil {
+		includePatternsSlice = *includePatterns
+	}
+
+	searchArgs := search.SymbolsParameters{
+		CommitID:        commitID,
+		First:           limitOrDefault(first) + 1, // add 1 so we can determine PageInfo.hasNextPage
+		Repo:            repoName.Name,
+		IncludePatterns: includePatternsSlice,
+	}
+	if query != nil {
+		searchArgs.Query = *query
+	}
+
+	symbols, err := backend.Symbols.ListTags(ctx, searchArgs)
+	if err != nil {
+		return nil, err
+	}
+
+	fileWithPath := func(path string) *result.File {
+		return &result.File{
+			Path:     path,
+			Repo:     repoName,
+			InputRev: inputRev,
+			CommitID: commitID,
+		}
+	}
+
+	matches := make([]*result.SymbolMatch, 0, len(symbols))
+	for _, symbol := range symbols {
+		matches = append(matches, &result.SymbolMatch{
+			Symbol: symbol,
+			File:   fileWithPath(symbol.Path),
+		})
+	}
+	return matches, err
+}
+
+const DefaultSymbolLimit = 100
+
+func limitOrDefault(first *int32) int {
+	if first == nil {
+		return DefaultSymbolLimit
+	}
+	return int(*first)
 }
